requirements.txt
requests
MLB-StatsAPI
pandas
tabulate
anvil-uplink
jsbeautifier